fix: 예약 결과 화면 회전시 여러번 저장되는 문제 해결
- 예약 결과 화면에서 저장을 진행하며 발생하는 문제
- 예약 확인 버튼을 눌렀을 때 예약 관련 이벤트(데이터 추가, 알림 설정)가 진행되도록 수정
