+++ app/src/main/java/woowacourse/movie/ui/seat/SeatSelectActivity.kt
@@ -1,10 +1,13 @@
 package woowacourse.movie.ui.seat
 
+import android.app.AlarmManager
+import android.app.PendingIntent
 import android.content.Intent
 import android.os.Bundle
 import android.widget.TextView
 import com.example.domain.usecase.DiscountApplyUseCase
 import woowacourse.movie.R
+import woowacourse.movie.data.TicketsRepository
 import woowacourse.movie.model.SeatPositionState
 import woowacourse.movie.model.TicketOptState
 import woowacourse.movie.model.TicketsState
@@ -12,6 +15,7 @@ import woowacourse.movie.model.mapper.asDomain
 import woowacourse.movie.model.mapper.asPresentation
 import woowacourse.movie.ui.BackKeyActionBarActivity
 import woowacourse.movie.ui.DecimalFormatters
+import woowacourse.movie.ui.confirm.AlarmReceiver
 import woowacourse.movie.ui.confirm.ReservationConfirmActivity
 import woowacourse.movie.ui.customView.ConfirmView
 import woowacourse.movie.ui.reservation.MovieDetailActivity.Companion.KEY_TICKETS
@@ -19,6 +23,9 @@ import woowacourse.movie.util.getParcelableArrayListCompat
 import woowacourse.movie.util.getParcelableExtraCompat
 import woowacourse.movie.util.keyError
 import woowacourse.movie.util.showAskDialog
+import java.time.LocalDateTime
+import java.util.*
+import kotlin.collections.ArrayList
 
 class SeatSelectActivity : BackKeyActionBarActivity() {
     private val discountApplyUseCase = DiscountApplyUseCase()
@@ -71,14 +78,18 @@ class SeatSelectActivity : BackKeyActionBarActivity() {
             messageId = R.string.ask_really_reservation,
             negativeStringId = R.string.reservation_cancel,
             positiveStringId = R.string.reservation_complete
-        ) {
-            navigateReservationConfirmActivity(seats)
-        }
+        ) { reserveMovie(seats) }
     }
 
-    private fun navigateReservationConfirmActivity(seats: List<SeatPositionState>) {
+    private fun reserveMovie(seats: List<SeatPositionState>) {
+        val tickets: TicketsState = TicketsState.from(ticketOptState, seats)
+        TicketsRepository.addTicket(tickets)
+        navigateReservationConfirmActivity(tickets)
+        setNotification(tickets)
+    }
+
+    private fun navigateReservationConfirmActivity(tickets: TicketsState) {
         val intent = Intent(this, ReservationConfirmActivity::class.java)
-        val tickets = TicketsState.from(ticketOptState, seats)
         intent.putExtra(KEY_TICKETS, tickets)
         startActivity(intent)
     }
@@ -96,6 +107,35 @@ class SeatSelectActivity : BackKeyActionBarActivity() {
         moneyTextView.text = getString(
             R.string.discount_money,
             DecimalFormatters.convertToMoneyFormat(discountApplyMoney.asPresentation())
+        )
+    }
+
+    private fun setNotification(tickets: TicketsState) {
+        val tickets = tickets.copy(dateTime = LocalDateTime.of(0, 4, 26, 17, 47, 30))
+        val calendar: Calendar = Calendar.getInstance().apply {
+            set(
+                tickets.dateTime.year,
+                tickets.dateTime.monthValue - 1,
+                tickets.dateTime.dayOfMonth,
+                tickets.dateTime.hour,
+                tickets.dateTime.minute
+            )
+        }
+        val alarmManager: AlarmManager = getSystemService(ALARM_SERVICE) as AlarmManager
+        val alarmIntent =
+            Intent(this, AlarmReceiver::class.java).apply { putExtra("a", tickets) }.let { intent ->
+                PendingIntent.getBroadcast(
+                    this,
+                    tickets.hashCode(),
+                    intent,
+                    PendingIntent.FLAG_UPDATE_CURRENT + PendingIntent.FLAG_IMMUTABLE
+                )
+            }
+
+        alarmManager.setExactAndAllowWhileIdle(
+            AlarmManager.RTC_WAKEUP,
+            calendar.timeInMillis - 30 * 60 * 1000,
+            alarmIntent
         )
     }
